# Simplify walkDir(...) and improve performance by roughly 10x in larger repos