## Description

The performance of WalkDir is improved by replacing the recursive DFS with a stack based implementation (less memory in-use during peak exploration). In addition, some unnecessary flags and options were removed which allowed the code to be overall simplified. Much less rule matching is occurring with the same files returned. Finally, all the event emitter features were not needed so they were removed.

Some performance results on my linux host:

small repo:
- 649 files returned by walkDir()
- 9 .gitignore files
- walkDir() old vs new duration: 918 ms to 143 ms

medium sized repo:
- 9271 returned by walkDir()
- 35 .gitignore files
- walkDir() old vs new duration: 10,886 ms to 970ms

large sized repo:
- 197,900 files returned by walkDir()
- 929 .gitignore files
- walkDir() old vs new duration: 498,151 ms to 37,353 ms

## Checklist

- [x] The base branch of this PR is `dev`, rather than `main`
- [x] The relevant docs, if any, have been updated or created

## Testing

I did my performance testing with this simple test:

```
describe.only("perf exploration", () => {
  test("time walkdir", async () => {
    const options = {
      returnRelativePaths: true,
    };
    console.time('walkDir.continue')
    const results = await walkDir('/home/username/code/common/continue/', ide, options);
    console.timeEnd('walkDir.continue');
    expect(results.length).toBe(649)
  }, 10 * 60 * 1000);
});
```