@sestinj 2024-07-23
Going to merge, since everything looks to be working perfectly after pulling it down. Keep us in the loop with progress on workers—I know that a previous attempt at using them for calculating embeddings lead to trickiness with VS Code, so if you run into some kind of wall there it might be useful to compare notes